Split into submodules and fixed some needless borrowing
